+++ opal/mca/common/ucx/Makefile.am
@@ -14,7 +14,6 @@
 
 headers =                         \
         common_ucx.h              \
-        common_ucx_int.h          \
         common_ucx_wpool.h        \
         common_ucx_wpool_int.h
